app/routes/shared.py
from datetime import datetime, timedelta, timezone

import jwt
from flask import current_app


def create_access_token(user_id):
    payload = {
        "sub": user_id,
        "type": "access",
        "exp": datetime.now(timezone.utc) + timedelta(hours=6),
    }
    return jwt.encode(payload, current_app.config["JWT_SECRET_KEY"], algorithm="HS256")


def decode_token(token):
    try:
        payload = jwt.decode(
            token,
            current_app.config["JWT_SECRET_KEY"],
            algorithms=["HS256"],
        )
        return payload.get("sub")
    except jwt.PyJWTError:
        return None
