requirements.txt
Flask==3.0.3
pymongo==4.8.0
PyJWT==2.9.0
python-dotenv==1.0.1
Werkzeug==3.0.3
numpy==2.1.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
